requirements.txt
# Core ML and Data Processing
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Progress and Monitoring
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# File Processing
osrparse>=7.0.0  # For .osr replay file parsing

# Utilities
pyyaml>=6.0
jupyter>=1.0.0
ipywidgets>=8.0.0

# AI Model Dependencies
transformers>=4.30.0  # For transformer architecture components
accelerate>=0.20.0    # For distributed training and mixed precision
datasets>=2.13.0      # For dataset management
einops>=0.6.0         # For tensor operations

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Optimization and Hyperparameter Tuning
optuna>=3.2.0

# CLI and Logging
click>=8.1.0
rich>=13.0.0
loguru>=0.7.0

# Memory Profiling and Performance
memory-profiler>=0.60.0
psutil>=5.9.0

# Development Tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0